Add files via upload
Analysis on the features of most common Java XML parsers that could lead to vulnerabilities such as XXE and SSRF. During this research a new attack vector using the "file" schema was detected that introduced a new payload for exploiting XXE issues. 
The research was originally made on August 2019 and published internally on the Immunity's internal wiki. The post was published in Immunity's Blog on February 7 of 2021 and it was nominee for the Top 10 Web Hacking Techniques of 2021.